requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
asttokens==2.0.8
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==0.10.0
fonttools==4.34.4
frozenlist==1.3.0
fsspec==2022.5.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.47.0
hydra-core==1.2.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.11.1
kiwisolver==1.4.4
Markdown==3.4.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
multidict==6.0.2
nest-asyncio==1.5.5
numpy==1.23.1
oauthlib==3.2.0
omegaconf==2.2.2
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.6.5
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
stack-data==0.4.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.12.0
torch-tb-profiler==0.4.0
torchaudio==0.12.0
torchmetrics==0.9.2
torchvision==0.13.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing-extensions==4.3.0
urllib3==1.26.10
wcwidth==0.2.5
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.1
